Crate leo_interpreter

Modules

core_functions ğŸ”’

cursor ğŸ”’

cursor_aleo ğŸ”’

dialoguer_input ğŸ”’

interpreter ğŸ”’

ratatui_ui ğŸ”’

ui ğŸ”’

util ğŸ”’

value ğŸ”’

Macros

halt

halt_no_span

tc_fail

Enums

Value

A Leo value of any type.

Constants

HELP ğŸ”’

INTRO ğŸ”’

Traits

CoreFunctionHelper

A context in which we can evaluate core functions.

Functions

evaluate_binary

Evaluate a binary operation.

evaluate_core_function

evaluate_unary

Evaluate a unary operation.

interpret

Load all the Leo source files indicated and open the interpreter to commands from the user.

literal_to_value

parse_breakpoint ğŸ”’

tokenize_user_input ğŸ”’