leo_interpreter
Module cursor

Structs

ContextStack

A stack of contexts, building with the function call stack.

Cursor

Tracks the current execution state - a cursor into the running program.

Frame

A frame of execution, keeping track of the Element next to be executed and the number of steps we’ve done so far.

FunctionContext

Names associated to values in a function being executed.

GlobalId

Global values - such as mappings, functions, etc - are identified by program and name.

StepResult

Enums

AleoContext

Element

A Leo construct to be evauated.

FunctionVariant

Functions

evaluate_binary

Evaluate a binary operation.

evaluate_unary

Evaluate a unary operation.

literal_to_value

Type Aliases

Closure

Finalize

SvmFunction