leo_interpreter::cursor

Struct Cursor

pub struct Cursor {

    pub frames: Vec<Frame>,
    pub values: Vec<Value>,
    pub functions: HashMap<Location, FunctionVariant>,
    pub async_blocks: HashMap<NodeID, Block>,
    pub globals: HashMap<Location, Value>,
    pub user_values: HashMap<Vec<Symbol>, Value>,
    pub mappings: HashMap<Location, HashMap<Value, Value>>,
    pub structs: HashMap<Vec<Symbol>, IndexMap<Symbol, Type>>,
    pub records: HashMap<(Symbol, Vec<Symbol>), IndexMap<Symbol, Type>>,
    pub futures: Vec<AsyncExecution>,
    pub contexts: ContextStack,
    pub signer: Value,
    pub rng: ChaCha20Rng,
    pub block_height: u32,
    pub really_async: bool,
    pub program: Option<Symbol>,
    pub network: NetworkName,
    pub private_key: String,

}

Tracks the current execution state - a cursor into the running program.

Fields
frames: Vec<Frame>

Stack of execution frames, with the one currently to be executed on top.

values: Vec<Value>

Stack of values from evaluated expressions.

Each time an expression completes evaluation, a value is pushed here.

functions: HashMap<Location, FunctionVariant>

All functions (or transitions or inlines) in any program being interpreted.

async_blocks: HashMap<NodeID, Block>

All the async blocks encountered. We identify them by their NodeID.

globals: HashMap<Location, Value>

Consts are stored here.

user_values: HashMap<Vec<Symbol>, Value>mappings: HashMap<Location, HashMap<Value, Value>>structs: HashMap<Vec<Symbol>, IndexMap<Symbol, Type>>

For each struct type, we only need to remember the names of its members, in order.

records: HashMap<(Symbol, Vec<Symbol>), IndexMap<Symbol, Type>>

For each record type, we index by program name and path, and remember its members except owner.

futures: Vec<AsyncExecution>contexts: ContextStacksigner: Valuerng: ChaCha20Rngblock_height: u32really_async: boolprogram: Option<Symbol>network: NetworkNameprivate_key: String
Implementations

impl Cursor

pub fn new( really_async: bool, private_key: String, block_height: u32, network: NetworkName, ) -> Self

really_async indicates we should really delay execution of async function calls until the user runs them.

pub fn clear(&mut self)

fn set_place( new_value: Value, this_value: &mut Value, places: &mut dyn Iterator<Item = &Expression>, indices: &mut dyn Iterator<Item = Value>, ) -> Result<()>

pub fn assign( &mut self, value: Value, place: &Expression, indices: &mut dyn Iterator<Item = Value>, ) -> Result<()>

pub fn set_program(&mut self, program: &str)

pub fn current_program(&self) -> Option<Symbol>

pub fn increment_step(&mut self)

fn new_caller(&self) -> Value

fn pop_value(&mut self) -> Result<Value>

fn lookup(&self, name: &[Symbol]) -> Option<Value>

pub fn lookup_mapping( &self, program: Option<Symbol>, name: Symbol, ) -> Option<&HashMap<Value, Value>>

pub fn lookup_mapping_mut( &mut self, program: Option<Symbol>, name: Symbol, ) -> Option<&mut HashMap<Value, Value>>

fn lookup_function( &self, program: Symbol, name: &[Symbol], ) -> Option<FunctionVariant>

fn set_variable(&mut self, path: &[Symbol], value: Value)

pub fn whole_step(&mut self) -> Result<StepResult>

Execute the whole step of the current Element.

That is, perform a step, and then finish all statements and expressions that have been pushed, until we’re ready for the next step of the current Element (if there is one).

pub fn over(&mut self) -> Result<StepResult>

Step over the current Element.

That is, continue executing until the current Element is finished.

pub fn step_block( &mut self, block: &Block, function_body: bool, step: usize, ) -> bool

fn to_absolute_path(&self, name: &[Symbol]) -> Vec<Symbol>

Returns the full absolute path by prefixing name with the current module path. If no context is available, returns name as-is.

fn step_statement(&mut self, statement: &Statement, step: usize) -> Result<bool>

fn step_expression( &mut self, expression: &Expression, expected_ty: &Option<Type>, step: usize, ) -> Result<bool>

pub fn step(&mut self) -> Result<StepResult>

Execute one step of the current element.

Many Leo constructs require multiple steps. For instance, when executing a conditional, the first step will push the condition expression to the stack. Once that has executed and we’ve returned to the conditional, we push the then or otherwise block to the stack. Once that has executed and we’ve returned to the conditional, the final step does nothing.

pub fn do_call( &mut self, function_program: Symbol, function_path: &[Symbol], arguments: impl Iterator<Item = Value>, finalize: bool, span: Span, ) -> Result<()>

impl Cursor

fn mapping_by_call_operator( &self, operator: &CallOperator<TestnetV0>, ) -> Option<&HashMap<Value, Value>>

fn get_register(&self, reg: &Register<TestnetV0>) -> Value

fn set_register(&mut self, reg: Register<TestnetV0>, value: Value)

fn instructions_len(&self) -> usize

fn increment_instruction_index(&mut self)

fn execution_complete(&self) -> bool

fn next_instruction(&self) -> Option<&Instruction<TestnetV0>>

fn next_command(&self) -> Option<&Command<TestnetV0>>

fn operand_value(&self, operand: &Operand<TestnetV0>) -> Value

fn step_aleo_instruction( &mut self, instruction: Instruction<TestnetV0>, ) -> Result<()>

fn outputs(&self) -> Vec<Value>

fn step_aleo_command(&mut self, command: Command<TestnetV0>) -> Result<()>

fn branch(&mut self, label: &Identifier<TestnetV0>)

pub fn step_aleo(&mut self) -> Result<()>