Crate leo_interpreter

Modules

cursor ğŸ”’

cursor_aleo ğŸ”’

dialoguer_input ğŸ”’

interpreter ğŸ”’

ratatui_ui ğŸ”’

ui ğŸ”’

util ğŸ”’

Macros

halt

halt_no_span

tc_fail

Structs

TestFunction

Constants

HELP ğŸ”’

INTRO ğŸ”’

Functions

find_and_run_tests

Run interpreter tests and return data about native tests.

interpret

Load all the Leo source files indicated and open the interpreter to commands from the user.

parse_breakpoint ğŸ”’

tokenize_user_input ğŸ”’