leo_interpreter

Module cursor

Structs

ContextStack

A stack of contexts, building with the function call stack.

Cursor

Tracks the current execution state - a cursor into the running program.

Frame

A frame of execution, keeping track of the Element next to be executed and the number of steps we’ve done so far.

FunctionContext

Names associated to values in a function being executed.

StepResult

Enums

AleoContext

Element

A Leo construct to be evauated.

FunctionVariant

Type Aliases

Closure

Finalize

SvmFunction